+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # SequenceDiagram Changelog
 
 ## [Unreleased]
+### Fixed
+- Export image width issue.
+- Change `Logger` to `com.intellij.openapi.diagnostic.Logger`
 
 ## [2.2.1]
 ### Added